fix: address PR review comments for LlmConfig
- Mark api_key as SensitiveField in spec and add to sensitive fields table
- Add missing url and api_key fields to reference sheet
- Clarify url field defaults to API provider's default URL when null
- Split adapter-specific tests into respective tests/adapters/ folders
